# packages/libs/registry-sdk/src/generated/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vincent Registry API",
    "version": "1.0.3",
    "description": "API for Vincent App, Tool, and Policy Registry"
  },
  "servers": [
    {
      "url": "https://staging.registry.heyvincent.ai"
    },
    {
      "url": "https://registry.heyvincent.ai"
    },
    {
      "url": "http://localhost:3000/"
    }
  ],
  "components": {
    "securitySchemes": {
      "siweAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Sign In With Ethereum authentication (SIWE). Format is \"SIWE [JSON({ message, signature })]\""
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid input provided"
          }
        },
        "required": [
          "message"
        ]
      },
      "DeleteResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Success message",
            "example": "Successfully deleted"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "AppVersionEdit": {
        "type": "object",
        "properties": {
          "changes": {
            "type": "string",
            "description": "Describes what changed between this version and the previous version.",
            "example": "I am a changelog trapped in a computer!"
          }
        },
        "additionalProperties": false
      },
...
      "AppList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/App"
        }
      },
      "AppVersionList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/AppVersion"
        }
      },
      "AppVersionToolList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/AppVersionTool"
        }
      },
...
    },
...
  },
  "paths": {
    "/apps": {
      "get": {
        "tags": [
          "App"
        ],
        "summary": "Lists all applications",
        "operationId": "listApps",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppList"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/app": {
      "post": {
        "tags": [
          "App",
          "AppVersion"
        ],
        "summary": "Creates a new application",
        "operationId": "createApp",
        "security": [
          {
            "siweAuth": []
          }
        ],
        "requestBody": {
          "description": "Developer-defined application information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/App"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/app/{appId}": {
      "get": {
        "tags": [
          "App"
        ],
        "summary": "Fetches an application",
        "operationId": "getApp",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/App"
                }
              }
            }
          },
          "404": {
            "description": "Application not found"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "App"
        ],
        "summary": "Edits an application",
        "operationId": "editApp",
        "security": [
          {
            "siweAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          }
        ],
        "requestBody": {
          "description": "Developer-defined updated application details",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/App"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "App",
          "AppVersion",
          "AppVersionTool"
        ],
        "summary": "Deletes an application",
        "operationId": "deleteApp",
        "security": [
          {
            "siweAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Resource successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/app/{appId}/undelete": {
      "post": {
        "tags": [
          "App",
          "AppVersion",
          "AppVersionTool"
        ],
        "summary": "Undeletes an application",
        "operationId": "undeleteApp",
        "security": [
          {
            "siweAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Resource successfully undeleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/app/{appId}/versions": {
      "get": {
        "tags": [
          "AppVersion"
        ],
        "summary": "Fetches all versions of an application",
        "operationId": "getAppVersions",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersionList"
                }
              }
            }
          },
          "404": {
            "description": "Application not found"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/app/{appId}/version/{version}": {
      "get": {
        "tags": [
          "AppVersion"
        ],
        "summary": "Fetches an application version",
        "operationId": "getAppVersion",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          },
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "Version # of the target application version",
            "example": 3,
            "name": "version",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersion"
                }
              }
            }
          },
          "404": {
            "description": "Application not found"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AppVersion"
        ],
        "summary": "Edits an application version",
        "operationId": "editAppVersion",
        "security": [
          {
            "siweAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          },
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "Version # of the target application version",
            "example": 3,
            "name": "version",
            "in": "path"
          }
        ],
        "requestBody": {
          "description": "Update version changes field",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppVersionEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersion"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AppVersion"
        ],
        "summary": "Deletes an application version",
        "operationId": "deleteAppVersion",
        "security": [
          {
            "siweAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          },
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "Version # of the target application version",
            "example": 3,
            "name": "version",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "OK - Resource successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Application or version not found"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/app/{appId}/version/{version}/enable": {
      "post": {
        "tags": [
          "AppVersion"
        ],
        "summary": "Enables an application version",
        "operationId": "enableAppVersion",
        "security": [
          {
            "siweAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          },
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "Version # of the target application version",
            "example": 3,
            "name": "version",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersion"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/app/{appId}/version/{version}/disable": {
      "post": {
        "tags": [
          "AppVersion"
        ],
        "summary": "Disables an application version",
        "operationId": "disableAppVersion",
        "security": [
          {
            "siweAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          },
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "Version # of the target application version",
            "example": 3,
            "name": "version",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersion"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "422": {
            "description": "Validation exception"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/app/{appId}/version/{version}/tools": {
      "get": {
        "tags": [
          "AppVersionTool"
        ],
        "summary": "Lists all tools for an application version",
        "operationId": "listAppVersionTools",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "ID of the target application",
            "example": 132,
            "name": "appId",
            "in": "path"
          },
          {
            "schema": {
              "type": "number"
            },
            "required": true,
            "description": "Version # of the target application version",
            "example": 3,
            "name": "version",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppVersionToolList"
                }
              }
            }
          },
          "404": {
            "description": "Application or version not found"
          },
          "default": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}